This commit was generated by cvs2svn to compensate for changes in r5817, which included commits to RCS files with non-trunk default branches.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@5818 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/cmdline.c
@@ -74,7 +74,7 @@ int cmdline_register_options(const cmdline_option_t *c)
 static cmdline_option_t *lookup(const char *name, int *is_ambiguous)
 {
     cmdline_option_t *match;
-    int name_len = strlen(name);
+    size_t name_len = strlen(name);
     int i;
 
     match = NULL;
